requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
altgraph @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/altgraph-0.17.2-py2.py3-none-any.whl
apiron==7.0.0
async-timeout==4.0.2
attrs==22.2.0
blurhash==1.1.4
certifi==2022.12.7
charset-normalizer==2.1.1
decorator==5.1.1
frozenlist==1.3.3
future @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/future-0.18.2-py3-none-any.whl
idna==3.4
macholib @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/macholib-1.15.2-py2.py3-none-any.whl
Mastodon.py==1.8.0
multidict==6.0.4
openai==0.26.4
openaiapi==0.0.2
python-dateutil==2.8.2
python-magic==0.4.27
requests==2.28.2
six @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/six-1.15.0-py2.py3-none-any.whl
tqdm==4.64.1
urllib3==1.26.14
yarl==1.8.2
